new SyncStart xmlrpc command for remote syncstarting, soundcard playback modules imported from MuSE
git-svn-id: svn://dyne.org/montevideo/ivysync@71 949728d9-16ea-0310-a75c-cbdf8430a4b8

+++ branches/lydia/decoder.h
@@ -89,8 +89,8 @@ class Decoder : public Thread, public Entry {
   int   getpos();
   void  setpos(int pos);
 
-  off64_t getoffset();
-  void setoffset(off64_t pos);
+  int  getoffset();
+  void setoffset(int pos);
 
   /** state flags for use in the inner loop
       the following booleans are changed by asynchronous calls
